Crate gfx_backend_gl

OpenGL implementation of a device, striving to support OpenGL 2.0 with at least VAOs, but using newer extensions when available.

Structs

Device
GL device.

GlContext

Info
OpenGL implementation information

Instance

PhysicalDevice

PlatformName
A unique platform identifier that does not change between releases

QueueFamily

Starc
Single-threaded Arc. Wrapper for Arc that allows you to Send it even if T: !Sync. Yet internal data cannot be accessed outside of the thread where it was created.

Surface

Swapchain

Version
A version number for a specific component of an OpenGL implementation

Wstarc
Single-threaded Weak. Wrapper for Weak that allows you to Send it even if T: !Sync. Yet internal data cannot be accessed outside of the thread where it was created.

Enums

Backend

Error